requirements.txt
numpy==1.26.4
PyQt5==5.15.6
SpeechRecognition==3.10.4
markdown==3.7
pynput==1.7.7
llama-cpp-python==0.3.1
huggingface_hub==0.25.1
openwakeword==0.6.0
pyinstaller==6.10.0
ffmpeg-python==0.2.0
llama-index-core==0.12.0
llama-index-embeddings-huggingface==0.4.0
docx2txt==0.8
git+https://github.com/stlukey/whispercpp.py